refactor: improve feedback modal ui
