+++ ui/src/app/components/ai-chat/ai-chat.component.html
@@ -136,6 +136,7 @@
                   <!-- Unified feedback badge with two buttons -->
                   <div
                     class="absolute -bottom-3 right-2 flex items-center bg-white rounded-full shadow-sm border border-secondary-200 p-0.5 z-10"
+                    *ngIf="showFeedbackBadge"
                   >
                     <button
                       class="text-xs text-success-600 p-1.5 rounded-full flex items-center transition-colors duration-200 hover:bg-secondary-50"
@@ -209,24 +210,26 @@
     *ngIf="isFeedbackModalOpen"
     class="fixed inset-0 bg-black bg-opacity-50 flex items-center justify-center z-50"
   >
-    <div class="bg-white rounded-lg p-6 w-96">
-      <h3 class="text-lg font-medium mb-4">
-        {{ feedbackType === "like" ? "Like" : "Dislike" }} Feedback
+    <div class="bg-white rounded-lg p-6 w-96 max-w-md">
+      <h3 class="text-base font-medium mb-4">
+        {{
+          feedbackType === "like" ? "Share what you liked" : "Help us improve"
+        }}
       </h3>
       <textarea
         [(ngModel)]="feedbackText"
-        placeholder="Optional feedback"
-        class="w-full h-24 p-2 border border-secondary-200 rounded-lg"
+        placeholder="Your thoughts help us enhance the experience (optional)"
+        class="w-full h-28 p-3 text-sm border border-secondary-200 rounded-lg"
       ></textarea>
-      <div class="flex justify-end gap-2 mt-4">
+      <div class="flex justify-end gap-3 mt-4">
         <button
-          class="bg-secondary-200 text-secondary-700 px-4 py-2 rounded-lg"
+          class="bg-secondary-200 text-secondary-700 px-4 py-2 text-sm rounded-lg"
           (click)="closeFeedbackModal()"
         >
           Cancel
         </button>
         <button
-          class="bg-primary-600 text-white px-4 py-2 rounded-lg"
+          class="bg-primary-600 text-white px-4 py-2 text-sm rounded-lg"
           (click)="submitFeedback()"
         >
           Submit
