+++ ui/src/app/services/analytics/utils/analytics.utils.ts
@@ -1,4 +1,7 @@
+import { BehaviorSubject } from 'rxjs';
+
 export const ANALYTICS_TOGGLE_KEY = 'analyticsEnabled';
+export const analyticsEnabledSubject = new BehaviorSubject<boolean>(getAnalyticsToggleState());
 
 export function getAnalyticsToggleState(): boolean {
   const storedValue = localStorage.getItem(ANALYTICS_TOGGLE_KEY);
@@ -7,4 +10,5 @@ export function getAnalyticsToggleState(): boolean {
 
 export function setAnalyticsToggleState(enabled: boolean): void {
   localStorage.setItem(ANALYTICS_TOGGLE_KEY, JSON.stringify(enabled));
+  analyticsEnabledSubject.next(enabled); 
 }
